Rename default and default light themes
