+++ packages/cli/src/ui/themes/default-light.ts
