+++ packages/cli/src/ui/themes/default.ts
