+++ packages/cli/src/ui/themes/theme-manager.ts
@@ -8,8 +8,8 @@ import { AtomOneDark } from './atom-one-dark.js';
 import { Dracula } from './dracula.js';
 import { GitHub } from './github.js';
 import { GoogleCode } from './googlecode.js';
-import { DefaultLight } from './vs.js';
-import { DefaultDark } from './vs2015.js';
+import { DefaultLight } from './default-light.js';
+import { DefaultDark } from './default.js';
 import { XCode } from './xcode.js';
 import { Theme, ThemeType } from './theme.js';
 import { ANSI } from './ansi.js';
